requirements.txt
streamlit
openai-whisper
keybert
sentence-transformers
numpy
pandas
scikit-learn
plotly